+++ app/models/organization.rb
@@ -5,6 +5,7 @@ class Organization < ApplicationRecord
   include OrganizationTimezone
   include Currencies
   include Organizations::AuthenticationMethods
+  include HasFeatureFlags
 
   self.ignored_columns += [:clickhouse_aggregation]
 
@@ -237,39 +238,6 @@ class Organization < ApplicationRecord
   # This is added to have a common interface for all organization-related models to access the organization.
   def organization
     self
-  end
-
-  # Feature flag methods
-  def feature_flag_enabled?(flag)
-    flag = flag.to_s
-    FeatureFlag.validate!(flag)
-
-    return false unless FeatureFlag.valid?(flag)
-
-    feature_flags.include?(flag)
-  end
-
-  def feature_flag_disabled?(flag)
-    flag = flag.to_s
-    !feature_flag_enabled?(flag)
-  end
-
-  def enable_feature_flag!(flag)
-    flag = flag.to_s
-    FeatureFlag.validate!(flag)
-
-    return unless FeatureFlag.valid?(flag)
-
-    update!(feature_flags: feature_flags | [flag]) unless feature_flags.include?(flag)
-  end
-
-  def disable_feature_flag!(flag)
-    flag = flag.to_s
-    FeatureFlag.validate!(flag)
-
-    return unless FeatureFlag.valid?(flag)
-
-    update!(feature_flags: feature_flags - [flag]) if feature_flags.include?(flag)
   end
 
   private
